+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/seekbar/Fingerprints.kt
@@ -1,6 +1,7 @@
 package app.revanced.patches.youtube.layout.seekbar
 
 import app.revanced.patcher.fingerprint
+import app.revanced.patches.youtube.shared.YOUTUBE_MAIN_ACTIVITY_CLASS_TYPE
 import app.revanced.util.containsLiteralInstruction
 import app.revanced.util.getReference
 import app.revanced.util.indexOfFirstInstruction
@@ -103,7 +104,7 @@ internal val launchScreenLayoutTypeFingerprint = fingerprint {
     custom { method, _ ->
         val firstParameter = method.parameterTypes.firstOrNull()
         // 19.25 - 19.45
-        (firstParameter == "Lcom/google/android/apps/youtube/app/watchwhile/MainActivity;"
+        (firstParameter == YOUTUBE_MAIN_ACTIVITY_CLASS_TYPE
                 || firstParameter == "Landroid/app/Activity;") // 19.46+
                 && method.containsLiteralInstruction(launchScreenLayoutTypeLotteFeatureFlag)
     }