+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/shortsautoplay/ShortsAutoplayPatch.kt
@@ -68,7 +68,7 @@ val shortsAutoplayPatch = bytecodePatch(
 
         // Main activity is used to check if app is in pip mode.
         mainActivityOnCreateFingerprint.method.addInstruction(
-            1,
+            0,
             "invoke-static/range { p0 .. p0 }, $EXTENSION_CLASS_DESCRIPTOR->setMainActivity(Landroid/app/Activity;)V",
         )