+++ patches/src/main/kotlin/app/revanced/patches/youtube/misc/extension/SharedExtensionPatch.kt
@@ -3,4 +3,5 @@ package app.revanced.patches.youtube.misc.extension
 import app.revanced.patches.shared.misc.extension.sharedExtensionPatch
 import app.revanced.patches.youtube.misc.extension.hooks.*
 
-val sharedExtensionPatch = sharedExtensionPatch("youtube", applicationInitHook)
+val sharedExtensionPatch = sharedExtensionPatch("youtube",
+    applicationInitHook, applicationInitOnCrateHook)